Created new classes for testing, also checking if my commits is registered
