+++ src/test/java/modelTest/characterSelectTest.java
@@ -1,4 +1,5 @@
 package modelTest;
 
 public class characterSelectTest {
+
 }
